backend-py/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
python-dotenv==1.1.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
